accessibility/20130226.html

<!DOCTYPE html>
<title>IRC logs: mozilla / #accessibility / 20130226</title>
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="mozilla/accessibility.20130226.log"><a href="/irc-logs/">/irc-logs</a> / mozilla / #accessibility / 2013-02-26 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Tue Feb 26 00:00:00 2013</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #accessibility</li>
 <li id="l-3"><a href="#l-3">#</a> [00:04] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-4" _a="@davidb"><a href="#l-4">#</a> [00:10] &lt;@davidb&gt; ha. ha.</li>
 <li id="l-5"><a href="#l-5">#</a> [00:24] * Quits: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net) (Ping timeout)</li>
 <li id="l-6"><a href="#l-6">#</a> [00:25] * Joins: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:28] * Joins: avcpp (bzlandfill@D46DA8FB.ADE68EA9.64781E5D.IP)</li>
 <li id="l-8" _a="@firebot"><a href="#l-8">#</a> [00:28] &lt;@firebot&gt; New Core - Disability Access APIs bug 845095 filed by philringnalda@gmail.com.</li>
 <li id="l-9" _a="@firebot"><a href="#l-9">#</a> [00:29] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=845095">https://bugzilla.mozilla.org/show_bug.cgi?id=845095</a> nor, --, ---, nobody, NEW, Intermittent Mac states/test_tree.xul | Test timed out. followed by states/test_visibility.html fail</li>
 <li id="l-10" _a="avcpp" _r="Hi"><a href="#l-10">#</a> [00:29] &lt;avcpp&gt; Hi, I'm working on but 429990 and after reviewing for a while I don't exactly understand what is being requested. Should the out arguments be changed to null at the end of the method so that they are null upon return of the method? Or should they all be initialized to null in the beginning of the method?</li>
 <li id="l-11" _a="@davidb" _r="tbsaunde"><a href="#l-11">#</a> [00:29] &lt;@davidb&gt; tbsaunde: I think you followed that bug? ^</li>
 <li id="l-12" _a="@tbsaunde" _r="avcpp"><a href="#l-12">#</a> [00:31] &lt;@tbsaunde&gt; avcpp: you should make sure that the pointers passed into the method are not null, and if they are just return E_INVALIDARG</li>
 <li id="l-13" _a="avcpp"><a href="#l-13">#</a> [00:32] &lt;avcpp&gt; ahh i see, thanks :) is this only for out arguments? or all pointers?</li>
 <li id="l-14" _a="@tbsaunde" _r="avcpp"><a href="#l-14">#</a> [00:35] &lt;@tbsaunde&gt; avcpp: all I guess, but iirc there almost all out args anyway</li>
 <li id="l-15" _a="avcpp"><a href="#l-15">#</a> [00:36] &lt;avcpp&gt; ok. One more question, I know that some files were being renamed and moved from the msaa folder to the windows directory, I was told that they would all be moved in time so only to focus on those in the windows folder. Should I only focus on the files in the ia2 folder of the windows directory?</li>
 <li id="l-16" _a="@tbsaunde" _r="avcpp"><a href="#l-16">#</a> [00:37] &lt;@tbsaunde&gt; avcpp: no, do sdn/ and uia/ too though there's much less in them</li>
 <li id="l-17" _a="avcpp" _r="tbsaunde"><a href="#l-17">#</a> [00:38] &lt;avcpp&gt; tbsaunde: ok so all in the windows directory. should i ignore the files still in msaa for now?</li>
 <li id="l-18" _a="lizzard"><a href="#l-18">#</a> [00:39] &lt;lizzard&gt; this is what i like to see when I introduce people :D</li>
 <li id="l-19"><a href="#l-19">#</a> [00:41] * Quits: habber (habber@moz-8654C0B7.nyc.res.rr.com) (Quit: habber)</li>
 <li id="l-20" _a="@tbsaunde" _r="avcpp"><a href="#l-20">#</a> [00:43] &lt;@tbsaunde&gt; avcpp: yes</li>
 <li id="l-21" _a="avcpp" _r="ok"><a href="#l-21">#</a> [00:43] &lt;avcpp&gt; ok, thank you very much</li>
 <li id="l-22" _a="@tbsaunde"><a href="#l-22">#</a> [00:44] &lt;@tbsaunde&gt; np</li>
 <li id="l-23"><a href="#l-23">#</a> [00:50] * Joins: fxa90id (fxa90id@moz-7923A087.dsl.dynamic.t-mobile.pl)</li>
 <li id="l-24" _a="@tbsaunde" _r="davidb"><a href="#l-24">#</a> [00:58] &lt;@tbsaunde&gt; davidb: fwiw I'm making sure my abomination of a patch builds</li>
 <li id="l-25" _a="@davidb" _r="tbsaunde"><a href="#l-25">#</a> [00:59] &lt;@davidb&gt; tbsaunde: cool</li>
 <li id="l-26" _a="@davidb" _r="tbsaunde"><a href="#l-26">#</a> [00:59] &lt;@davidb&gt; tbsaunde: on try?</li>
 <li id="l-27" _a="@tbsaunde" _r="davidb"><a href="#l-27">#</a> [01:00] &lt;@tbsaunde&gt; davidb: no, locally (mingw)</li>
 <li id="l-28" _a="@davidb"><a href="#l-28">#</a> [01:00] &lt;@davidb&gt; heh ok</li>
 <li id="l-29" _a="avcpp" _r="Hello"><a href="#l-29">#</a> [01:01] &lt;avcpp&gt; Hello, I realized that in some cases the pointers are set to null, for instance in ia2AccessibleAction.cpp. Should I ignore these cases?</li>
 <li id="l-30" _a="@tbsaunde" _r="avcpp"><a href="#l-30">#</a> [01:06] &lt;@tbsaunde&gt; avcpp: you mean *aOutPtr = null?</li>
 <li id="l-31" _a="@tbsaunde" _r="avcpp"><a href="#l-31">#</a> [01:07] &lt;@tbsaunde&gt; avcpp: you should check the pointer before we set it to point at null</li>
 <li id="l-32" _a="avcpp" _r="tbsaunde"><a href="#l-32">#</a> [01:09] &lt;avcpp&gt; tbsaunde: the specific line I was looking at was *aDescription = NULL, or *ppv = NULL</li>
 <li id="l-33" _a="khuey"><a href="#l-33">#</a> [01:10] * khuey is now known as khuey|away</li>
 <li id="l-34" _a="@tbsaunde" _r="avcpp"><a href="#l-34">#</a> [01:10] &lt;@tbsaunde&gt; avcpp: yeah, then check the pointer before assigning to it</li>
 <li id="l-35" _a="avcpp" _r="ok"><a href="#l-35">#</a> [01:11] &lt;avcpp&gt; ok, thanks</li>
 <li id="l-36"><a href="#l-36">#</a> [01:14] * Quits: avcpp (bzlandfill@D46DA8FB.ADE68EA9.64781E5D.IP) (Quit: CGI:IRC (EOF))</li>
 <li id="l-37"><a href="#l-37">#</a> [01:42] * Quits: @davidb (davidb@C593061F.9FC507A8.7B545C66.IP) (Quit: davidb)</li>
 <li id="l-38"><a href="#l-38">#</a> [02:04] * Quits: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net) (Input/output error)</li>
 <li id="l-39"><a href="#l-39">#</a> [02:05] * Joins: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net)</li>
 <li id="l-40" _a="@firebot"><a href="#l-40">#</a> [02:09] &lt;@firebot&gt; New Core - Disability Access APIs bug 845134 filed by ryanvm@gmail.com.</li>
 <li id="l-41" _a="@firebot"><a href="#l-41">#</a> [02:09] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=845134">https://bugzilla.mozilla.org/show_bug.cgi?id=845134</a> nor, --, ---, nobody, NEW, Intermittent content/a11y/accessible/focus/test_focusedChild.html | Test timed out.</li>
 <li id="l-42"><a href="#l-42">#</a> [02:24] * Quits: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net) (Quit: nhirata)</li>
 <li id="l-43"><a href="#l-43">#</a> [03:23] * Quits: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net) (Quit: scott_gonzalez)</li>
 <li id="l-44" _a="khuey|away"><a href="#l-44">#</a> [03:34] * khuey|away is now known as khuey</li>
 <li id="l-45"><a href="#l-45">#</a> [03:51] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-46"><a href="#l-46">#</a> [04:02] * Joins: surkov (surkov@7F1C0E2B.41FE424A.E17943EE.IP)</li>
 <li id="l-47" _a="ChanServ"><a href="#l-47">#</a> [04:02] * ChanServ sets mode: +o surkov</li>
 <li id="l-48" _a="@firebot"><a href="#l-48">#</a> [04:11] &lt;@firebot&gt; trev.saunders@gmail.com granted review for attachment 716942 on bug 527461.</li>
 <li id="l-49" _a="@firebot"><a href="#l-49">#</a> [04:11] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=527461">https://bugzilla.mozilla.org/show_bug.cgi?id=527461</a> enh, --, ---, surkov.alexander, ASSI, Implement RELATION_NODE_PARENT_OF</li>
 <li id="l-50" _a="@surkov" _r="tbsaunde"><a href="#l-50">#</a> [04:20] &lt;@surkov&gt; tbsaunde: hi</li>
 <li id="l-51" _a="@tbsaunde" _r="surkov"><a href="#l-51">#</a> [04:20] &lt;@tbsaunde&gt; surkov: so, I was looking at the xul label thing, any particular reason you don't just use normal text leaf and get rid of the special class?</li>
 <li id="l-52" _a="@tbsaunde"><a href="#l-52">#</a> [04:21] &lt;@tbsaunde&gt; it doesn't seem like it should be hard to do</li>
 <li id="l-53" _a="@surkov"><a href="#l-53">#</a> [04:21] &lt;@surkov&gt; don't we have any assumptions about frame type of text leafs?</li>
 <li id="l-54" _a="@surkov" _r="tbsaunde"><a href="#l-54">#</a> [04:22] &lt;@surkov&gt; tbsaunde: does it look ok:</li>
 <li id="l-55" _a="@surkov"><a href="#l-55">#</a> [04:22] &lt;@surkov&gt; MOZ_STATIC_ASSERT(relationTypes[ATK_RELATION_CONTROLLED_BY + 1] == nsIAccessibleRelation::RELATION_CONTROLLED_BY,</li>
 <li id="l-56" _a="@surkov"><a href="#l-56">#</a> [04:22] &lt;@surkov&gt;                     &quot;RELATION_CONTROLLED_BY is not in sync!&quot;);</li>
 <li id="l-57" _a="@tbsaunde" _r="surkov"><a href="#l-57">#</a> [04:22] &lt;@tbsaunde&gt; surkov: yes</li>
 <li id="l-58" _a="@tbsaunde" _r="surkov"><a href="#l-58">#</a> [04:23] &lt;@tbsaunde&gt; surkov: I don't think text leaf has any dependancy on frame time from quick look at code</li>
 <li id="l-59" _a="@surkov" _r="tbsaunde"><a href="#l-59">#</a> [04:23] &lt;@surkov&gt; tbsaunde: I could take a look, actually I just copied/pasted bullet thing</li>
 <li id="l-60" _a="@tbsaunde" _r="surkov"><a href="#l-60">#</a> [04:24] &lt;@tbsaunde&gt; surkov: I see</li>
 <li id="l-61" _a="@tbsaunde" _r="surkov"><a href="#l-61">#</a> [04:40] &lt;@tbsaunde&gt; surkov: anything else I should look at tonight?</li>
 <li id="l-62" _a="@surkov" _r="tbsaunde"><a href="#l-62">#</a> [04:41] &lt;@surkov&gt; tbsaunde: might be not</li>
 <li id="l-63"><a href="#l-63">#</a> [05:07] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-64"><a href="#l-64">#</a> [05:11] * Joins: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net)</li>
 <li id="l-65" _a="@firebot"><a href="#l-65">#</a> [05:14] &lt;@firebot&gt; New Core - Disability Access APIs bug 845176 filed by philringnalda@gmail.com.</li>
 <li id="l-66" _a="@firebot"><a href="#l-66">#</a> [05:14] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=845176">https://bugzilla.mozilla.org/show_bug.cgi?id=845176</a> nor, --, ---, nobody, NEW, Intermittent relations/test_embeds.xul | Test timed out.</li>
 <li id="l-67"><a href="#l-67">#</a> [05:33] * Quits: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net) (Quit: scott_gonzalez)</li>
 <li id="l-68"><a href="#l-68">#</a> [05:52] * Joins: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net)</li>
 <li id="l-69"><a href="#l-69">#</a> [05:53] * Quits: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net) (Quit: scott_gonzalez)</li>
 <li id="l-70"><a href="#l-70">#</a> [05:53] * Joins: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net)</li>
 <li id="l-71"><a href="#l-71">#</a> [06:04] * Quits: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net) (Quit: scott_gonzalez)</li>
 <li id="l-72"><a href="#l-72">#</a> [06:35] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-73"><a href="#l-73">#</a> [06:38] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Ping timeout)</li>
 <li id="l-74"><a href="#l-74">#</a> [06:38] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-75"><a href="#l-75">#</a> [06:51] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-76"><a href="#l-76">#</a> [06:52] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-77" _a="@tbsaunde" _r="surkov"><a href="#l-77">#</a> [06:56] &lt;@tbsaunde&gt; surkov: the array thing is sad making, but ok</li>
 <li id="l-78" _a="@surkov"><a href="#l-78">#</a> [06:57] &lt;@surkov&gt; ok</li>
 <li id="l-79" _a="@surkov" _r="tbsaunde"><a href="#l-79">#</a> [06:57] &lt;@surkov&gt; tbsaunde: so I land without those, ok?</li>
 <li id="l-80" _a="@tbsaunde" _r="surkov"><a href="#l-80">#</a> [07:00] &lt;@tbsaunde&gt; surkov: yeah</li>
 <li id="l-81" _a="@surkov"><a href="#l-81">#</a> [07:00] &lt;@surkov&gt; k</li>
 <li id="l-82" _a="@tbsaunde" _r="surkov"><a href="#l-82">#</a> [07:01] &lt;@tbsaunde&gt; surkov: I guess you might try moving the array out of the function since function static and file static are somewhat different</li>
 <li id="l-83" _a="@surkov"><a href="#l-83">#</a> [07:02] &lt;@surkov&gt; I tried with same result</li>
 <li id="l-84" _a="@tbsaunde" _r="surkov"><a href="#l-84">#</a> [07:13] &lt;@tbsaunde&gt; surkov: ok, no worries good to try</li>
 <li id="l-85"><a href="#l-85">#</a> [07:25] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-86"><a href="#l-86">#</a> [07:33] * Quits: peteb-away (ptbrunet@moz-E9B02845.austin.res.rr.com) (Ping timeout)</li>
 <li id="l-87"><a href="#l-87">#</a> [07:39] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-88"><a href="#l-88">#</a> [07:42] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Quit: ChatZilla 0.9.90 [Firefox 22.0a1/20130225031141])</li>
 <li id="l-89" _a="khuey"><a href="#l-89">#</a> [08:00] * khuey is now known as khuey|away</li>
 <li id="l-90" _a="@firebot"><a href="#l-90">#</a> [08:13] &lt;@firebot&gt; surkov.alexander@gmail.com granted in-testsuite on bug 527461.</li>
 <li id="l-91" _a="@firebot"><a href="#l-91">#</a> [08:13] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=527461">https://bugzilla.mozilla.org/show_bug.cgi?id=527461</a> enh, --, ---, surkov.alexander, ASSI, Implement RELATION_NODE_PARENT_OF</li>
 <li id="l-92"><a href="#l-92">#</a> [08:33] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-93"><a href="#l-93">#</a> [08:36] * Quits: fxa90id (fxa90id@moz-7923A087.dsl.dynamic.t-mobile.pl) (Quit: Leaving)</li>
 <li id="l-94"><a href="#l-94">#</a> [08:36] * Quits: @surkov (surkov@7F1C0E2B.41FE424A.E17943EE.IP) (Quit: surkov)</li>
 <li id="l-95"><a href="#l-95">#</a> [08:50] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Ping timeout)</li>
 <li id="l-96"><a href="#l-96">#</a> [08:52] * Quits: mgorse (mgorse@B6C068B2.2088E15E.A18F112A.IP) (Ping timeout)</li>
 <li id="l-97"><a href="#l-97">#</a> [08:53] * Joins: mgorse (mgorse@B6C068B2.2088E15E.A18F112A.IP)</li>
 <li id="l-98" _a="@firebot"><a href="#l-98">#</a> [09:41] &lt;@firebot&gt; ioana.budnar@softvision.ro changed the Component on bug 842576 from Untriaged to Keyboard Navigation.</li>
 <li id="l-99" _a="@firebot"><a href="#l-99">#</a> [09:41] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=842576">https://bugzilla.mozilla.org/show_bug.cgi?id=842576</a> nor, --, ---, nobody, UNCO, [quick find on a page] '/' search shortcut works only with English keyboard input</li>
 <li id="l-100"><a href="#l-100">#</a> [10:30] * Quits: victorporof (victorporo@1CF03390.183F338A.6A4F8DA2.IP) (Quit: victorporof)</li>
 <li id="l-101"><a href="#l-101">#</a> [11:05] * Quits: mgorse (mgorse@B6C068B2.2088E15E.A18F112A.IP) (Ping timeout)</li>
 <li id="l-102"><a href="#l-102">#</a> [11:08] * Joins: mgorse (mgorse@B6C068B2.2088E15E.A18F112A.IP)</li>
 <li id="l-103"><a href="#l-103">#</a> [13:06] * Joins: amolpol98 (Mibbit@7061CD01.D6BD0D91.22D41759.IP)</li>
 <li id="l-104" _a="amolpol98"><a href="#l-104">#</a> [13:06] &lt;amolpol98&gt; hii</li>
 <li id="l-105" _a="amolpol98"><a href="#l-105">#</a> [13:07] &lt;amolpol98&gt; I am a newbie. Can anyone help me with how to start fixing bug for mozilla</li>
 <li id="l-106"><a href="#l-106">#</a> [13:09] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-107"><a href="#l-107">#</a> [13:11] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Quit: ChatZilla 0.9.90 [Firefox 22.0a1/20130225031141])</li>
 <li id="l-108"><a href="#l-108">#</a> [13:35] * Joins: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net)</li>
 <li id="l-109"><a href="#l-109">#</a> [13:36] * Parts: amolpol98 (Mibbit@7061CD01.D6BD0D91.22D41759.IP)</li>
 <li id="l-110"><a href="#l-110">#</a> [14:07] * Joins: habber (habber@BC5C67E2.155E3EA2.67CD788C.IP)</li>
 <li id="l-111"><a href="#l-111">#</a> [14:08] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-112"><a href="#l-112">#</a> [14:13] * Joins: peteb-away (ptbrunet@moz-E9B02845.austin.res.rr.com)</li>
 <li id="l-113"><a href="#l-113">#</a> [14:15] * Joins: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-114"><a href="#l-114">#</a> [14:25] * Quits: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net) (Quit: scott_gonzalez)</li>
 <li id="l-115"><a href="#l-115">#</a> [14:31] * Joins: scott_gonzalez (scott_gonz@moz-91C81A39.hrbgpa.fios.verizon.net)</li>
 <li id="l-116" _a="khuey|away"><a href="#l-116">#</a> [14:58] * khuey|away is now known as khuey</li>
 <li id="l-117"><a href="#l-117">#</a> [14:58] * Quits: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP) (Quit: Justin_o)</li>
 <li id="l-118"><a href="#l-118">#</a> [15:02] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Ping timeout)</li>
 <li id="l-119"><a href="#l-119">#</a> [15:25] * Joins: habber_ (habber@moz-8654C0B7.nyc.res.rr.com)</li>
 <li id="l-120"><a href="#l-120">#</a> [15:27] * Quits: habber (habber@BC5C67E2.155E3EA2.67CD788C.IP) (Ping timeout)</li>
 <li id="l-121" _a="habber_"><a href="#l-121">#</a> [15:27] * habber_ is now known as habber</li>
 <li id="l-122"><a href="#l-122">#</a> [15:27] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-123"><a href="#l-123">#</a> [15:31] * Quits: habber (habber@moz-8654C0B7.nyc.res.rr.com) (Ping timeout)</li>
 <li id="l-124"><a href="#l-124">#</a> [15:40] * Joins: habber (habber@BC5C67E2.155E3EA2.67CD788C.IP)</li>
 <li id="l-125"><a href="#l-125">#</a> [15:42] * Quits: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP) (Ping timeout)</li>
 <li id="l-126"><a href="#l-126">#</a> [16:05] * Joins: victorporof (victorporo@1CF03390.183F338A.6A4F8DA2.IP)</li>
 <li id="l-127"><a href="#l-127">#</a> [16:06] * Quits: mgorse (mgorse@B6C068B2.2088E15E.A18F112A.IP) (Quit: ircII EPIC4-2.2 -- Are we there yet?)</li>
 <li id="l-128"><a href="#l-128">#</a> [16:32] * Joins: clown (clown@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-129" _a="khuey"><a href="#l-129">#</a> [16:47] * khuey is now known as khuey|away</li>
 <li id="l-130"><a href="#l-130">#</a> [17:01] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-131" _a="@firebot"><a href="#l-131">#</a> [17:01] &lt;@firebot&gt; ryanvm@gmail.com changed the Resolution on bug 527461 from --- to FIXED.</li>
 <li id="l-132" _a="@firebot"><a href="#l-132">#</a> [17:02] &lt;@firebot&gt; ryanvm@gmail.com changed the Status on bug 527461 from ASSIGNED to RESOLVED.</li>
 <li id="l-133" _a="@firebot"><a href="#l-133">#</a> [17:02] &lt;@firebot&gt; ryanvm@gmail.com changed the Target Milestone on bug 527461 from --- to mozilla22.</li>
 <li id="l-134" _a="@firebot"><a href="#l-134">#</a> [17:02] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=527461">https://bugzilla.mozilla.org/show_bug.cgi?id=527461</a> enh, --, mozilla22, surkov.alexander, RESO FIXED, Implement RELATION_NODE_PARENT_OF</li>
 <li id="l-135" _a="khuey|away"><a href="#l-135">#</a> [17:04] * khuey|away is now known as khuey</li>
 <li id="l-136"><a href="#l-136">#</a> [17:18] * Joins: nhirata (anonymous@moz-BE33DA21.fw1.sfo1.mozilla.net)</li>
 <li id="l-137"><a href="#l-137">#</a> [17:21] * Joins: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-138"><a href="#l-138">#</a> [17:37] * Quits: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP) (Quit: Justin_o)</li>
 <li id="l-139"><a href="#l-139">#</a> [17:42] * Joins: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-140"><a href="#l-140">#</a> [17:46] * Joins: davidb (davidb@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-141" _a="ChanServ"><a href="#l-141">#</a> [17:46] * ChanServ sets mode: +qo davidb davidb</li>
 <li id="l-142" _a="@davidb"><a href="#l-142">#</a> [17:47] &lt;@davidb&gt; heyo</li>
 <li id="l-143" _a="@tbsaunde"><a href="#l-143">#</a> [17:48] &lt;@tbsaunde&gt; hey</li>
 <li id="l-144"><a href="#l-144">#</a> [17:59] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-145"><a href="#l-145">#</a> [18:29] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-146"><a href="#l-146">#</a> [18:36] * Quits: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP) (Quit: Justin_o)</li>
 <li id="l-147"><a href="#l-147">#</a> [18:57] * Joins: marcoz (marco.zehe@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-148" _a="ChanServ"><a href="#l-148">#</a> [18:58] * ChanServ sets mode: +o marcoz</li>
 <li id="l-149"><a href="#l-149">#</a> [18:59] * Joins: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-150" _a="@davidb" _r="eeejay"><a href="#l-150">#</a> [19:24] &lt;@davidb&gt; eeejay: bug 805856</li>
 <li id="l-151" _a="@firebot"><a href="#l-151">#</a> [19:24] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=805856">https://bugzilla.mozilla.org/show_bug.cgi?id=805856</a> nor, --, ---, continuation, NEW, Shutdown CC logging doesn't work any more</li>
 <li id="l-152" _a="@firebot"><a href="#l-152">#</a> [19:35] &lt;@firebot&gt; dbolter@mozilla.com changed the Resolution on bug 534558 from --- to INCOMPLETE.</li>
 <li id="l-153" _a="@firebot"><a href="#l-153">#</a> [19:35] &lt;@firebot&gt; dbolter@mozilla.com changed the Status on bug 534558 from NEW to RESOLVED.</li>
 <li id="l-154" _a="@firebot"><a href="#l-154">#</a> [19:35] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=534558">https://bugzilla.mozilla.org/show_bug.cgi?id=534558</a> nor, --, ---, nobody, RESO INCOMPLETE, ARIA shouldn't always override strong native semantics</li>
 <li id="l-155" _a="@tbsaunde" _r="davidb"><a href="#l-155">#</a> [20:00] &lt;@tbsaunde&gt; davidb: have you tried that patch yet?</li>
 <li id="l-156" _a="@davidb" _r="tbsaunde"><a href="#l-156">#</a> [20:00] &lt;@davidb&gt; tbsaunde: i completely forgot. i will try it today though</li>
 <li id="l-157" _a="@tbsaunde" _r="davidb"><a href="#l-157">#</a> [20:01] &lt;@tbsaunde&gt; davidb: ok, good luck</li>
 <li id="l-158" _a="@davidb"><a href="#l-158">#</a> [20:01] &lt;@davidb&gt; ty</li>
 <li id="l-159" _a="@tbsaunde" _r="heh"><a href="#l-159">#</a> [20:04] &lt;@tbsaunde&gt; heh, I sort of hope it doesn't help</li>
 <li id="l-160" _a="@firebot"><a href="#l-160">#</a> [20:04] &lt;@firebot&gt; hub@mozilla.com requested review from dbolter@mozilla.com  for attachment 718534 on bug 845095.</li>
 <li id="l-161" _a="@firebot"><a href="#l-161">#</a> [20:04] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=845095">https://bugzilla.mozilla.org/show_bug.cgi?id=845095</a> nor, --, ---, nobody, NEW, Intermittent Mac states/test_tree.xul | Test timed out. followed by states/test_visibility.html fail</li>
 <li id="l-162"><a href="#l-162">#</a> [20:05] * Quits: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP) (Quit: Justin_o)</li>
 <li id="l-163" _a="@tbsaunde" _r="hub"><a href="#l-163">#</a> [20:06] &lt;@tbsaunde&gt; hub: just check ^ in unless you feel like turning on logging to get logs to debug it</li>
 <li id="l-164" _a="@hub" _r="tbsaunde"><a href="#l-164">#</a> [20:06] &lt;@hub&gt; tbsaunde: r=you?</li>
 <li id="l-165" _a="@tbsaunde" _r="hub"><a href="#l-165">#</a> [20:07] &lt;@tbsaunde&gt; hub: rs=me if you like I guess</li>
 <li id="l-166" _a="@hub" _r="tbsaunde"><a href="#l-166">#</a> [20:07] &lt;@hub&gt; tbsaunde: I'll answer your review comment though</li>
 <li id="l-167" _a="@hub"><a href="#l-167">#</a> [20:07] &lt;@hub&gt; you might be right</li>
 <li id="l-168" _a="@tbsaunde" _r="hub"><a href="#l-168">#</a> [20:07] &lt;@tbsaunde&gt; hub: ?</li>
 <li id="l-169" _a="@hub"><a href="#l-169">#</a> [20:08] &lt;@hub&gt; &quot;I forget... do you need to call SimpleTest.finish() here?&quot;</li>
 <li id="l-170" _a="@hub"><a href="#l-170">#</a> [20:08] &lt;@hub&gt; oh wait</li>
 <li id="l-171" _a="@hub"><a href="#l-171">#</a> [20:08] &lt;@hub&gt; that's davidb</li>
 <li id="l-172" _a="@hub"><a href="#l-172">#</a> [20:08] &lt;@hub&gt; *sigh*</li>
 <li id="l-173" _a="@davidb"><a href="#l-173">#</a> [20:08] &lt;@davidb&gt; hi</li>
 <li id="l-174" _a="@davidb" _r="hub"><a href="#l-174">#</a> [20:09] &lt;@davidb&gt; hub: is it a silly question?</li>
 <li id="l-175" _a="@hub" _r="davidb"><a href="#l-175">#</a> [20:11] &lt;@hub&gt; davidb: no I think you are right</li>
 <li id="l-176" _a="@davidb"><a href="#l-176">#</a> [20:11] &lt;@davidb&gt; if so, we probably need it elsewhere as well</li>
 <li id="l-177" _a="@tbsaunde" _r="well"><a href="#l-177">#</a> [20:11] &lt;@tbsaunde&gt; well, if you call waitForExplicitFinish() I think you do, but I don't know why you wouldn't just do if (mac) { todo(&quot;mac this work on mac&quot;); return; }</li>
 <li id="l-178" _a="@firebot"><a href="#l-178">#</a> [20:11] &lt;@firebot&gt; dbolter@mozilla.com granted review for attachment 718534 on bug 845095.</li>
 <li id="l-179" _a="@davidb"><a href="#l-179">#</a> [20:12] &lt;@davidb&gt; I think what hub has is probably ok.</li>
 <li id="l-180" _a="@hub"><a href="#l-180">#</a> [20:13] &lt;@hub&gt; it is temporary</li>
 <li id="l-181" _a="@hub"><a href="#l-181">#</a> [20:13] &lt;@hub&gt; I'll need to fix it</li>
 <li id="l-182" _a="@davidb"><a href="#l-182">#</a> [20:13] * @davidb nods</li>
 <li id="l-183" _a="@hub"><a href="#l-183">#</a> [20:13] &lt;@hub&gt; and it is on 10.6 :-/</li>
 <li id="l-184" _a="@davidb"><a href="#l-184">#</a> [20:14] &lt;@davidb&gt; i really don't like test fragility</li>
 <li id="l-185" _a="@hub"><a href="#l-185">#</a> [20:14] &lt;@hub&gt; neither</li>
 <li id="l-186" _a="@davidb" _r="tbsaunde"><a href="#l-186">#</a> [20:15] &lt;@davidb&gt; tbsaunde: updating my tree in my windows vm… man hg pull is slow</li>
 <li id="l-187" _a="@davidb"><a href="#l-187">#</a> [20:16] &lt;@davidb&gt; luckily i have decent compilation strength now</li>
 <li id="l-188" _a="@davidb" _r="hub"><a href="#l-188">#</a> [20:19] &lt;@davidb&gt; hub: to clear i think your patch is fine as is.</li>
 <li id="l-189" _a="@davidb"><a href="#l-189">#</a> [20:19] &lt;@davidb&gt; +be</li>
 <li id="l-190" _a="@hub" _r="davidb"><a href="#l-190">#</a> [20:20] &lt;@hub&gt; davidb: I have a try run pending</li>
 <li id="l-191" _a="@hub"><a href="#l-191">#</a> [20:20] &lt;@hub&gt; just to make sure</li>
 <li id="l-192" _a="@davidb"><a href="#l-192">#</a> [20:20] &lt;@davidb&gt; ok</li>
 <li id="l-193"><a href="#l-193">#</a> [20:20] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-194"><a href="#l-194">#</a> [20:21] * Joins: SteveF (chatzilla@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-195"><a href="#l-195">#</a> [20:21] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-196" _a="@davidb"><a href="#l-196">#</a> [20:29] &lt;@davidb&gt; come on mercurial</li>
 <li id="l-197" _a="@davidb"><a href="#l-197">#</a> [20:38] &lt;@davidb&gt; my python is too old to run pymake now?</li>
 <li id="l-198" _a="@davidb"><a href="#l-198">#</a> [20:39] * @davidb downloads a new mozillabuild</li>
 <li id="l-199" _a="@tbsaunde" _r="hub"><a href="#l-199">#</a> [20:50] &lt;@tbsaunde&gt; hub: if its only  10.6 then running it on macosx !10.6 and dumping it at the back of the queue may well be reasonable</li>
 <li id="l-200" _a="@davidb"><a href="#l-200">#</a> [20:54] * @davidb clobbers</li>
 <li id="l-201"><a href="#l-201">#</a> [20:57] * Joins: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-202" _a="@davidb"><a href="#l-202">#</a> [20:57] &lt;@davidb&gt; yay for build scroll!</li>
 <li id="l-203" _a="khuey"><a href="#l-203">#</a> [20:59] * khuey is now known as khuey|away</li>
 <li id="l-204" _a="@davidb" _r="tbsaunde"><a href="#l-204">#</a> [21:03] &lt;@davidb&gt; tbsaunde: looking at your patch… it isn't strictly horrible</li>
 <li id="l-205" _a="@tbsaunde" _r="davidb"><a href="#l-205">#</a> [21:05] &lt;@tbsaunde&gt; davidb: oh?</li>
 <li id="l-206" _a="@davidb"><a href="#l-206">#</a> [21:05] &lt;@davidb&gt; yeah it is pretty clear what you are doing at least.</li>
 <li id="l-207" _a="@tbsaunde" _r="ok"><a href="#l-207">#</a> [21:06] &lt;@tbsaunde&gt; ok, sure the insanity is blatent</li>
 <li id="l-208" _a="@davidb"><a href="#l-208">#</a> [21:06] &lt;@davidb&gt; lol</li>
 <li id="l-209" _a="@hub"><a href="#l-209">#</a> [21:09] &lt;@hub&gt; looks like a lot of people are missing their flight to CSUN through LHR</li>
 <li id="l-210" _a="@davidb"><a href="#l-210">#</a> [21:10] &lt;@davidb&gt; crappy</li>
 <li id="l-211" _a="@davidb" _r="hub"><a href="#l-211">#</a> [21:10] &lt;@davidb&gt; hub: BTW you are needinfo'ed on bug 840612</li>
 <li id="l-212" _a="@firebot"><a href="#l-212">#</a> [21:11] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=840612">https://bugzilla.mozilla.org/show_bug.cgi?id=840612</a> cri, --, ---, vyang, NEW, Marionette turns off NetworkManager's control of offline state too aggressively</li>
 <li id="l-213" _a="@hub"><a href="#l-213">#</a> [21:11] &lt;@hub&gt; I saw</li>
 <li id="l-214" _a="@davidb"><a href="#l-214">#</a> [21:11] &lt;@davidb&gt; ok</li>
 <li id="l-215" _a="@hub" _r="davidb"><a href="#l-215">#</a> [21:12] &lt;@hub&gt; davidb: mm, I actually thought it had been cleared because somebody else said they were on it</li>
 <li id="l-216" _a="@davidb"><a href="#l-216">#</a> [21:14] &lt;@davidb&gt; i guess they left that to you</li>
 <li id="l-217"><a href="#l-217">#</a> [21:15] * Quits: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net) (Quit: lizzard)</li>
 <li id="l-218"><a href="#l-218">#</a> [21:26] * Joins: lizzard (ehenry@moz-5A4BE8E8.dsl.static.sonic.net)</li>
 <li id="l-219"><a href="#l-219">#</a> [21:58] * Joins: habber_ (habber@moz-8654C0B7.nyc.res.rr.com)</li>
 <li id="l-220"><a href="#l-220">#</a> [21:59] * Quits: habber (habber@BC5C67E2.155E3EA2.67CD788C.IP) (Ping timeout)</li>
 <li id="l-221" _a="habber_"><a href="#l-221">#</a> [21:59] * habber_ is now known as habber</li>
 <li id="l-222" _a="khuey|away"><a href="#l-222">#</a> [22:07] * khuey|away is now known as khuey</li>
 <li id="l-223"><a href="#l-223">#</a> [22:08] * Quits: Justin_o (Justin_o@67828CC7.C1A51174.9D42CF23.IP) (Quit: Justin_o)</li>
 <li id="l-224"><a href="#l-224">#</a> [22:40] * Quits: @davidb (davidb@C593061F.9FC507A8.7B545C66.IP) (Ping timeout)</li>
 <li id="l-225"><a href="#l-225">#</a> [22:40] * Quits: victorporof (victorporo@1CF03390.183F338A.6A4F8DA2.IP) (Quit: victorporof)</li>
 <li id="l-226"><a href="#l-226">#</a> [22:52] * Joins: satdav (satdav@moz-2B0F19ED.cable.virginmedia.com)</li>
 <li id="l-227" _a="satdav" _r="marcoz"><a href="#l-227">#</a> [22:53] &lt;satdav&gt; marcoz, wonder if you can help me out, would you be free any wednesday about midnight your time or 3pm pst</li>
 <li id="l-228" _a="satdav"><a href="#l-228">#</a> [22:53] &lt;satdav&gt; it was to speak to the sumo team about accessibility</li>
 <li id="l-229"><a href="#l-229">#</a> [22:59] * Quits: @marcoz (marco.zehe@C593061F.9FC507A8.7B545C66.IP) (Quit: Leaving.)</li>
 <li id="l-230"><a href="#l-230">#</a> [22:59] * Parts: clown (clown@67828CC7.C1A51174.9D42CF23.IP)</li>
 <li id="l-231" _a="@firebot"><a href="#l-231">#</a> [23:02] &lt;@firebot&gt; trev.saunders@gmail.com granted review for attachment 716577 on bug 843637.</li>
 <li id="l-232" _a="@firebot"><a href="#l-232">#</a> [23:03] &lt;@firebot&gt; Bug <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=843637">https://bugzilla.mozilla.org/show_bug.cgi?id=843637</a> nor, --, ---, surkov.alexander, ASSI, move MSAA folder under Windows folder</li>
 <li id="l-233" _a="satdav" _r="tbsaunde"><a href="#l-233">#</a> [23:03] &lt;satdav&gt; tbsaunde, would you be free one wednesday about 3pm pst</li>
 <li id="l-234" _a="@tbsaunde" _r="satdav"><a href="#l-234">#</a> [23:04] &lt;@tbsaunde&gt; satdav: probably not that's dinner time</li>
 <li id="l-235" _a="satdav"><a href="#l-235">#</a> [23:04] &lt;satdav&gt; thats fine</li>
 <li id="l-236" _a="satdav" _r="tbsaunde"><a href="#l-236">#</a> [23:05] &lt;satdav&gt; tbsaunde, what time you done your dinner at</li>
 <li id="l-237" _a="satdav"><a href="#l-237">#</a> [23:05] &lt;satdav&gt; as I am sure we can scedule it around you</li>
 <li id="l-238" _a="satdav"><a href="#l-238">#</a> [23:05] &lt;satdav&gt; as would like someone to come and speak to us at sumo about accessibility</li>
 <li id="l-239" _a="@tbsaunde" _r="satdav"><a href="#l-239">#</a> [23:06] &lt;@tbsaunde&gt; satdav: depends, besides I'm not the person you want for that</li>
 <li id="l-240" _a="satdav" _r="tbsaunde"><a href="#l-240">#</a> [23:07] &lt;satdav&gt; tbsaunde, it would be good as you from the end user side</li>
 <li id="l-241" _a="satdav"><a href="#l-241">#</a> [23:07] &lt;satdav&gt; as we are speaking about the mobile site</li>
 <li id="l-242" _a="satdav"><a href="#l-242">#</a> [23:07] &lt;satdav&gt; also for sumo</li>
 <li id="l-243" _a="@tbsaunde" _r="satdav"><a href="#l-243">#</a> [23:08] &lt;@tbsaunde&gt; satdav: I'm ... not a good end user for your purposes, I have never  used sumo and don't really have much interest in needing to</li>
 <li id="l-244" _a="satdav" _r="tbsaunde"><a href="#l-244">#</a> [23:10] &lt;satdav&gt; tbsaunde, who would be the best to speak to</li>
 <li id="l-245" _a="satdav"><a href="#l-245">#</a> [23:10] &lt;satdav&gt; at qa</li>
 <li id="l-246"><a href="#l-246">#</a> [23:10] * Joins: victorporof (victorporo@1CF03390.183F338A.6A4F8DA2.IP)</li>
 <li id="l-247" _a="@tbsaunde"><a href="#l-247">#</a> [23:11] &lt;@tbsaunde&gt; no clue I don't do qa</li>
 <li id="l-248" _a="satdav"><a href="#l-248">#</a> [23:13] &lt;satdav&gt; well at accessibility</li>
 <li id="l-249"><a href="#l-249">#</a> [23:17] * Quits: victorporof (victorporo@1CF03390.183F338A.6A4F8DA2.IP) (Ping timeout)</li>
 <li id="l-250"><a href="#l-250">#</a> [23:44] * Joins: marcoz (marco.zehe@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-251"><a href="#l-251">#</a> [23:46] * Joins: davidb (davidb@C593061F.9FC507A8.7B545C66.IP)</li>
 <li id="l-252" _a="ChanServ"><a href="#l-252">#</a> [23:46] * ChanServ sets mode: +qo davidb davidb</li>
 <li id="l-253" _a="ChanServ"><a href="#l-253">#</a> [23:47] * ChanServ sets mode: +o marcoz</li>
 <li id="l-254"><a href="#l-254">#</a> Session Close: Wed Feb 27 00:00:00 2013</li>
</ol>
<p id="end">The end :)
